refactor(dashboard): extract pure char-list parsers from DevServer
Move 7 stateless parsing helpers (formatObjectTypeHex, parseCharListNumber/Boolean/ObjectTypes, parseDashboardEquipmentTokens, buildDashboardUniqueItemLookup, decodeDashboardEnchantIds) plus the DashboardAccountEquipmentToken type out of the DevServer god-class into src/dashboard/server/charListParsers.ts. None used instance state (this). DevServer.ts: 5,340 -> 5,253 lines. tsc clean — all 35 call sites rewired.
